backend/requirements.txt
asgiref==3.2.7
boto3==1.13.5
botocore==1.16.5
Django==3.0.6
django-cors-headers==3.2.1
django-filter==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
